Move Carto::Http::Request class to separate file #3778
